# Implement some operations of R3Element using intrinsics