--- a/geometry/r3_element_body.hpp
+++ b/geometry/r3_element_body.hpp
@@ -181,16 +181,10 @@ R3Element<Scalar> operator-(R3Element<Scalar> const& right) {
 template<typename Scalar>
 R3Element<Scalar> operator+(R3Element<Scalar> const& left,
                             R3Element<Scalar> const& right) {
-  __m128d const* const right_xy = reinterpret_cast<__m128d const*>(&right.x);
-  __m128d const* const left_xy = reinterpret_cast<__m128d const*>(&left.x);
-  __m128d const result_xy = _mm_add_pd(*left_xy, *right_xy);
-  Scalar const* const result_x =
-      reinterpret_cast<Scalar const*>(&result_xy.m128d_f64[0]);
-  Scalar const* const result_y =
-      reinterpret_cast<Scalar const*>(&result_xy.m128d_f64[1]);
-  return R3Element<Scalar>(*result_x,
-                           *result_y,
-                           left.z + right.z);
+  R3Element<Scalar> result;
+  result.xy = _mm_add_pd(left.xy, right.xy);
+  result.zt = _mm_add_sd(left.zt, right.zt);
+  return result;
 }
 
 template<typename Scalar>
@@ -228,18 +222,12 @@ R3Element<Scalar> operator/(R3Element<Scalar> const& left,
 template<typename LDimension, typename RScalar>
 R3Element<Product<Quantity<LDimension>, RScalar>>
 operator*(Quantity<LDimension> const& left, R3Element<RScalar> const& right) {
-  using Scalar = Product<Quantity<LDimension>, RScalar>;
-  __m128d const* const right_xy = reinterpret_cast<__m128d const*>(&right.x);
-  double const* const left_d = reinterpret_cast<double const*>(&left);
-  __m128d const left_xy = {*left_d, *left_d};
-  __m128d const result_xy = _mm_mul_pd(left_xy, *right_xy);
-  Scalar const* const result_x =
-      reinterpret_cast<Scalar const*>(&result_xy.m128d_f64[0]);
-  Scalar const* const result_y =
-      reinterpret_cast<Scalar const*>(&result_xy.m128d_f64[1]);
-  return R3Element<Scalar>(*result_x,
-                           *result_y,
-                           left * right.z);
+  double const* const left_double = reinterpret_cast<double const*>(&left);
+  __m128d const left_128d = _mm_load1_pd(left_double);
+  R3Element<Product<Quantity<LDimension>, RScalar>> result;
+  result.xy = _mm_mul_pd(left_128d, right.xy);
+  result.zt = _mm_mul_sd(left_128d, right.zt);
+  return result;
 }
 
 template<typename LScalar, typename RDimension>