--- a/geometry/r3_element_body.hpp
+++ b/geometry/r3_element_body.hpp
@@ -118,12 +118,12 @@ Square<Scalar> R3Element<Scalar>::Norm²() const {
   // changing it, use IACA to check the effect.
   __m128d const zero = _mm_setzero_pd();
   __m128d const z_0 = _mm_unpacklo_pd(zt, zero);
-  __m128d const z2_0 = _mm_mul_sd(z_0, z_0);
-  __m128d const x2_y2 = _mm_mul_pd(xy, xy);
-  __m128d const x2z2_y2 = _mm_add_pd(x2_y2, z2_0);
-  __m128d const result_0 = _mm_hadd_pd(x2z2_y2, zero);
+  __m128d const z²_0 = _mm_mul_sd(z_0, z_0);
+  __m128d const x²_y² = _mm_mul_pd(xy, xy);
+  __m128d const x²z²_y² = _mm_add_sd(x²_y², z²_0);
+  __m128d const x²y²z²_0 = _mm_hadd_pd(x²z²_y², zero);
   Square<Scalar> const* const result =
-      reinterpret_cast<Square<Scalar> const*>(&result_0.m128d_f64[0]);
+      reinterpret_cast<Square<Scalar> const*>(&x²y²z²_0.m128d_f64[0]);
   return *result;
 #else
   return x * x + y * y + z * z;
@@ -405,8 +405,7 @@ Product<LScalar, RScalar> Dot(R3Element<LScalar> const& left,
   __m128d const zero = _mm_setzero_pd();
   __m128d const lzrz = _mm_mul_sd(left.zt, right.zt);
   __m128d const lxrx_lyry = _mm_mul_pd(left.xy, right.xy);
-  __m128d const lzrz_0 = _mm_unpacklo_pd(lzrz, zero);
-  __m128d const lxrxlzrz_lyry = _mm_add_pd(lxrx_lyry, lzrz_0);
+  __m128d const lxrxlzrz_lyry = _mm_add_sd(lxrx_lyry, lzrz);
   __m128d const lxrxlyrylzrz_0 = _mm_hadd_pd(lxrxlzrz_lyry, zero);
   Product<LScalar, RScalar> const* const result =
       reinterpret_cast<Product<LScalar, RScalar> const*>(