… geometry/r3_element_test.cpp
… geometry/rotation_test.cpp
… numerics/polynomial_test.cpp
… principia.props
… quantities/quantities.hpp
… quantities/quantities_body.hpp