According to IACA the throughput of these operations goes from 3.00 to 2.00 cycles on Haswell, and this is reflected by a gain of up to 40% on the benchmarks of the polynomials.  As expected, the effect is more pronounced for Estrin evaluation.

Before:
```
------------------------------------------------------------------------------------------------------------
Benchmark                                                                     Time           CPU Iterations
------------------------------------------------------------------------------------------------------------
BM_EvaluatePolynomialInMonomialBasisDisplacement<EstrinEvaluator>/4        5502 ns       5616 ns     100000
BM_EvaluatePolynomialInMonomialBasisDisplacement<EstrinEvaluator>/8       10087 ns      10013 ns      74786
BM_EvaluatePolynomialInMonomialBasisDisplacement<EstrinEvaluator>/12      13518 ns      13455 ns      49857
BM_EvaluatePolynomialInMonomialBasisDisplacement<EstrinEvaluator>/16      19715 ns      19608 ns      37393
BM_EvaluatePolynomialInMonomialBasisDisplacement<HornerEvaluator>/4        5124 ns       5148 ns     100000
BM_EvaluatePolynomialInMonomialBasisDisplacement<HornerEvaluator>/8       10750 ns      10708 ns      64102
BM_EvaluatePolynomialInMonomialBasisDisplacement<HornerEvaluator>/12      16643 ns      16062 ns      40792
BM_EvaluatePolynomialInMonomialBasisDisplacement<HornerEvaluator>/16      24517 ns      24475 ns      28045
```
After:
```
------------------------------------------------------------------------------------------------------------
Benchmark                                                                     Time           CPU Iterations
------------------------------------------------------------------------------------------------------------
BM_EvaluatePolynomialInMonomialBasisDisplacement<EstrinEvaluator>/4        4041 ns       4068 ns     172583
BM_EvaluatePolynomialInMonomialBasisDisplacement<EstrinEvaluator>/8        7448 ns       7509 ns     112179
BM_EvaluatePolynomialInMonomialBasisDisplacement<EstrinEvaluator>/12       9890 ns      10013 ns      74786
BM_EvaluatePolynomialInMonomialBasisDisplacement<EstrinEvaluator>/16      13873 ns      13907 ns      56089
BM_EvaluatePolynomialInMonomialBasisDisplacement<HornerEvaluator>/4        4393 ns       4381 ns     160255
BM_EvaluatePolynomialInMonomialBasisDisplacement<HornerEvaluator>/8        9637 ns       9804 ns      74786
BM_EvaluatePolynomialInMonomialBasisDisplacement<HornerEvaluator>/12      15384 ns      15332 ns      49857
BM_EvaluatePolynomialInMonomialBasisDisplacement<HornerEvaluator>/16      22898 ns      22424 ns      29914
```