@aw1621107 2018-02-15
Clang on macOS can't find `<intrin.h>`. Including `<x86intrin.h>` seems to do the trick, though.