@eggrobin 2018-02-16
We should probably just use `<emmintrin.h>`, see https://stackoverflow.com/questions/11228855/header-files-for-x86-simd-intrinsics.