@pleroy 2018-02-17
There is no `x86intrin.h` on MSVC.  I am going for `nmmintrin.h` (n as in Nehalem) which hopefully exists everywhere.  This gives us all of SSE and none of AVX, which is what we want at this point.